Refactor platform logic into seperate file.  Ensure spec depends on building/copying the bin
